Create CardContent Component that renders month, year and streak
